requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
asyncio-atexit==1.0.1
attrs==25.3.0
autogen-agentchat==0.7.4
autogen-core==0.7.4
autogen-ext==0.7.4
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
azure-identity==1.24.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.5
bleach==6.2.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
cobble==0.1.4
coloredlogs==15.0.1
cryptography==45.0.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docker==7.1.0
docopt==0.6.2
durationpy==0.10
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2024.12.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.12.3
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kubernetes==33.1.0
langchain==0.3.27
langchain-core==0.3.75
langchain-neo4j==0.5.0
langchain-text-splitters==0.3.11
langsmith==0.4.21
loguru==0.7.3
lxml==6.0.1
magika==0.6.2
mammoth==1.10.0
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.3
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.13.1
mdurl==0.1.2
mem0ai==0.1.116
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
neo4j==5.28.2
neo4j-graphrag==1.9.1
numpy==2.3.2
oauthlib==3.3.1
olefile==0.47
ollama==0.5.3
onnxruntime==1.22.1
openai==1.102.0
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
pdfminer.six==20250506
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.4.0
playwright==1.55.0
portalocker==3.2.0
posthog==5.4.0
prompt_toolkit==3.0.52
protobuf==5.29.5
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pynvim==0.5.2
pypdf==5.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
qdrant-client==1.15.1
rank-bm25==0.2.2
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scipy==1.16.1
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
SQLAlchemy==2.0.43
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tinycss2==1.4.0
tokenizers==0.22.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.17.3
types-PyYAML==6.0.12.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
xlrd==2.0.2
xlsxwriter==3.2.5
yarg==0.1.9
youtube-transcript-api==1.0.3
zipp==3.23.0
zstandard==0.24.0
